ArgParse: modernize, no raw new/delete, and make header-only.
